requirements.txt
lightning
torchvision
wandb
gradio
timm
jsonargparse[signatures]
kaggle
pillow
rich
numpy
albumentations
